requirements.txt
astroid==2.2.5
certifi==2019.9.11
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.2.0
psycopg2==2.8.4
pylint==2.3.1
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wincertstore==0.2
wrapt==1.11.2
